Number_guesser_CODSOFT1
This C++ program is an interactive game that challenges the user's guessing skills. It generates a random number between a specified range and asks the user to guess it. The user will enter their guess, and the program will provide feedback on whether the guess is too high or too low. The game will continue until the user correctly guesses the random number.

The program uses concepts of random number generation, user input, and conditional statements to create an engaging and enjoyable guessing experience. Players can keep guessing until they successfully determine the hidden number, adding an element of excitement and suspense to the game.