Simple_caluclator_CODSOFT2
This C++ program is designed to create a simple calculator that allows users to perform basic arithmetic operations such as addition, subtraction, multiplication, and division on two numbers of their choice. The program provides a user-friendly interface that prompts the user to input two numerical values and select an operation from the menu. Once the user makes their choice, the program calculates the result and displays it to the user.

Key Features:
1-Input two numerical values from the user.
2-Present a menu for selecting an arithmetic operation.
3-Perform the chosen operation on the provided numbers.
4-Display the result to the user.
5-Allow the user to perform additional calculations or exit the program.
 It showcases basic C++ programming concepts, including user input, conditional statements, and arithmetic operations.